more sacrifices to the volcano god I mean linter

+++ Code/CPP/euler/PrimeNumbers.cpp
@@ -1,27 +1,24 @@
-#include "PrimeNumbers.h"
+#include "PrimeNumbers.h"    // NOLINT(build/include)
 
 //----------------
-void PrimeNumbers::addPrimeNumber()
-{
+void PrimeNumbers::addPrimeNumber() {
   if (!primeNumbers.size()) {
     primeNumbers.push_back(1L);
-  }
-  else {
-    unsigned long endp = primeNumbers[ primeNumbers.size() - 1 ];
-    unsigned long newp = nextPrime( endp );
-    primeNumbers.push_back( newp );
+  } else {
+    uint64_t endp = primeNumbers[ primeNumbers.size() - 1 ];
+    uint64_t newp = nextPrime(endp);
+    primeNumbers.push_back(newp);
   }
 }
 
 
 //----------------
-bool PrimeNumbers::isPrime(unsigned long num) const
-{
+bool PrimeNumbers::isPrime(uint64_t num) const {
   if (1L == num) return true;
   if (2L == num) return true;
   if (3L == num) return true;
 
-  unsigned long m = num - 1L;
+  uint64_t m = num - 1L;
   while (m > 1L) {
     if (0L == num % m) return false;
     --m;
@@ -32,24 +29,21 @@ bool PrimeNumbers::isPrime(unsigned long num) const
 
 
 //----------------
-unsigned long PrimeNumbers::nextPrime(unsigned long val) const
-{
-  unsigned long next = val + 1L;
+uint64_t PrimeNumbers::nextPrime(uint64_t val) const {
+  uint64_t next = val + 1L;
   while ( !isPrime(next) ) ++next;
   return next;
 }
 
 
 //----------------
-unsigned PrimeNumbers::size() const 
-{
+unsigned PrimeNumbers::size() const {
   return primeNumbers.size();
 }
 
 
 //----------------
-unsigned long PrimeNumbers::operator[](unsigned i) const 
-{
+uint64_t PrimeNumbers::operator[](uint64_t i) const {
   return primeNumbers[i];
 }
